+++ .pylintdict
@@ -10,7 +10,6 @@ codeword
 contravariance
 currentmodule
 cvs
-cvxpy
 data's
 datapoints
 dataset
