--- docs/apidocs/qiskit_machine_learning.multiclass_extensions.rst
@@ -1,6 +1,0 @@
-.. _qiskit-machine-learning-multiclass_extensions:
-
-.. automodule:: qiskit_machine_learning.multiclass_extensions
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
