--- docs/apidocs/qiskit_machine_learning.utils.rst
@@ -1,6 +1,0 @@
-.. _qiskit-machine-learning-utils:
-
-.. automodule:: qiskit_machine_learning.utils
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
