+++ qiskit_machine_learning/algorithms/distribution_learners/numpy_discriminator.py
@@ -31,7 +31,7 @@ logger = logging.getLogger(__name__)
 # pylint: disable=invalid-name
 
 
-class DiscriminatorNet():
+class DiscriminatorNet:
     """
     Discriminator
