--- qiskit_machine_learning/multiclass_extensions/__init__.py
@@ -1,59 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2018, 2021.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""
-Multi-class Extensions (:mod:`qiskit_machine_learning.multiclass_extensions`)
-=============================================================================
-A multiclass extension allows Machine Learning's binary classifier algorithms, such as
-:class:`~qiskit_machine_learning.algorithms.QSVM` and
-:class:`~qiskit_machine_learning.algorithms.SklearnSVM`
-to handle more than two classes and do
-`multiclass classification <https://en.wikipedia.org/wiki/Multiclass_classification>`_.
-
-The multiclass extensions use different techniques to perform multiclass classification
-using the underlying binary classifier.
-
-.. currentmodule:: qiskit_machine_learning.multiclass_extensions
-
-Multiclass Extension Base Class
-===============================
-
-.. autosummary::
-   :toctree: ../stubs/
-   :nosignatures:
-
-   MulticlassExtension
-
-Multiclass Extensions
-=====================
-
-.. autosummary::
-   :toctree: ../stubs/
-   :nosignatures:
-
-   AllPairs
-   OneAgainstRest
-   ErrorCorrectingCode
-
-"""
-
-from .estimator import Estimator
-from .multiclass_extension import MulticlassExtension
-from .all_pairs import AllPairs
-from .one_against_rest import OneAgainstRest
-from .error_correcting_code import ErrorCorrectingCode
-
-__all__ = ['Estimator',
-           'MulticlassExtension',
-           'AllPairs',
-           'OneAgainstRest',
-           'ErrorCorrectingCode']
