+++ setup.py
@@ -69,7 +69,6 @@ setuptools.setup(
     python_requires=">=3.6",
     extras_require={
         'torch': ["torch"],
-        'cvx': ['cvxpy>1.0.0,!=1.1.0,!=1.1.1,!=1.1.2,!=1.1.8'],
     },
     zip_safe=False
 )
