+++ test/datasets/test_breast_cancer.py
@@ -12,6 +12,7 @@
 
 """ Test Breast Cancer """
 
+import unittest
 from test import QiskitMachineLearningTestCase
 import json
 import numpy as np
@@ -50,3 +51,7 @@ class TestBreastCancer(QiskitMachineLearningTestCase):
         for key, _ in test_input.items():
             np.testing.assert_allclose(test_input[key].tolist(), test_input_ref[key], rtol=1e-04)
         np.testing.assert_array_equal(class_labels, list(training_input.keys()))
+
+
+if __name__ == '__main__':
+    unittest.main()
