+++ test/datasets/test_wine.py
@@ -12,11 +12,11 @@
 
 """ Test Wine """
 
+import unittest
 from test import QiskitMachineLearningTestCase
 import json
 import numpy as np
-from qiskit_machine_learning.datasets import wine
-from qiskit_machine_learning.utils import split_dataset_to_data_and_labels
+from qiskit_machine_learning.datasets import wine, split_dataset_to_data_and_labels
 
 
 class TestWine(QiskitMachineLearningTestCase):
@@ -56,3 +56,7 @@ class TestWine(QiskitMachineLearningTestCase):
         datapoints, class_to_label = split_dataset_to_data_and_labels(test_input)
         np.testing.assert_array_equal(datapoints[1], [0, 0, 1, 1, 1, 2, 2, 2, 2, 2])
         self.assertDictEqual(class_to_label, {'A': 0, 'B': 1, 'C': 2})
+
+
+if __name__ == '__main__':
+    unittest.main()
